Calculation

+++ Martium.TravelInfo.App/Models/TravelInfoSettingsModel.cs
@@ -3,8 +3,8 @@
     public class TravelInfoSettingsModel
     {
         public string DepartureCountry { get; set; }
-        public decimal PricePerKm { get; set; }
-        public decimal AdditionalDistanceInKm { get; set; }
+        public double PricePerKm { get; set; }
+        public double AdditionalDistanceInKm { get; set; }
         public string DepartureAddress { get; set; }
     }
 }
